Refactor: change Library Item Types package location to 'constants'
