+++ src/com/twu/biblioteca/constants/LibraryItemTypes.java
@@ -1,4 +1,4 @@
-package com.twu.biblioteca.domain;
+package com.twu.biblioteca.constants;
 
 public class LibraryItemTypes {
     public static final String BOOK = "book";
