+++ src/com/twu/biblioteca/domain/Book.java
@@ -1,5 +1,7 @@
 package com.twu.biblioteca.domain;
 
+import com.twu.biblioteca.constants.LibraryItemTypes;
+
 import java.util.UUID;
 
 public class Book implements LibraryItem {
